table - move useStyles from main table index.js to a separate file
